refactor: apply review feedback for 404 page (#77)
