+++ src/pages/index.ts
@@ -1,2 +1,2 @@
 export { default as Main } from './Main'
-export { NotFoundPage } from './not-fouond'
+export { NotFoundPage } from './not-found'
